struct -> datatype, remove duplicate tests
